requirements.txt
miros~=4.2.1
pillow~=10.4.0